[standardize exception]
The name of the exceptions in php-webdriver is not quite good because of
a legacy reason.

The new php-webdriver was rewritten but it reused the WebDriverException
from the old version. This is bad to have non-standard exception names
in many
senses.

Here is a common situation, when the test throws an exception, we might
find
that it is very hard to search for the solution via search engines
because of the
odd exceptions name. However, if we search it again using the standard
name,
we get more solutions and probably fix the bug immediately and easily.

Try to google:
1) webdriver unknownserverexception
2) webdriver unhandledwebdrivererror

The first one gives 10,800 results and it is the standard name.
The second one gives 362 results and it is our bad name.

again 1:
1) StaleElementReferenceException <-- 6900 results
2) ObsoleteElementWebDriverError <-- 196 results

again 2
1) NoSuchElementWebDriverError <-- 194 results
2) NoSuchElementException <-- 324,000 results

There are many solutions or blog posts talking about different tricky or
common errors.
We could reuse those resource by just renaming the exceptions.
This will very beneficial in long term in terms of maintaining,
debugging php-webdriver issue,
for us and people using it outside facebook.