RemoteWebElement: allow overriding construction
Previously, if a developer wanted to extend the RemoteWebElement()
class, you couldn't effectively do it. The name of the class
was hard-coded in a private factory method of RemoteWebDriver
that couldn't be overriden.

Now, a custom class that extends RemoteWebDriver can
construct a class derived from RemoteWebElement, allowing
extra functionality to be added to Elements themselves.

The same logic applies to the EventFiringWebElement.